Added fallback for failed conversions to zga

+++ app/app.php
@@ -139,6 +139,8 @@ $app->post("/image", function () use ($app) {
                 $fileNames[ 8 ] = "zga_" . $metadata . "_" . $filePrefix . ".jpg";
                 $time = microtime(true) - $start;
                 $app['monolog']->addDebug("$time Calling montage");
+		$app['monolog']->addDebug(" timeout 25 montage " . $_FILES["imagefile"]["tmp_name"] . " -coalesce -tile x1111 -frame 0 -geometry '+0+0' -quality 80 -colors 256 -background none -bordercolor none /tmp/media/".$fileNames[ 8 ]);
+
                 $v = exec(  " timeout 25 montage " . $_FILES["imagefile"]["tmp_name"] . " -coalesce -tile x1111 -frame 0 -geometry '+0+0' -quality 80 -colors 256 -background none -bordercolor none /tmp/media/".$fileNames[ 8 ]);           
                 $time = microtime(true) - $start;
                 $app['monolog']->addDebug("$time Called montage");
@@ -146,17 +148,20 @@ $app->post("/image", function () use ($app) {
                 try {
                     $montage = new Imagick( "/tmp/media/".$fileNames[ 8 ]);
                 } catch ( ImagickException $e ) {
-                    return new Response("Invalid image",500);
+                    $useZga = false;
+                    //return new Response("Invalid image",500);
                 }
-
+                if ( isset($montage) ) {
         	    $files [ 8 ] = $montage->getImageBlob();
-                $montage->destroy();
+                    $montage->destroy();
                 
-                $time = microtime(true) - $start;
-                $app['monolog']->addDebug("$time Cleaned up montage stuff");
-                $zgaSize = filesize("/tmp/media/".$fileNames[ 8 ]);
-                $originalGifSize = filesize($_FILES["imagefile"]["tmp_name"]);
-                if ($originalGifSize < $zgaSize) {
+                    $time = microtime(true) - $start;
+                    $app['monolog']->addDebug("$time Cleaned up montage stuff");
+                    $zgaSize = filesize("/tmp/media/".$fileNames[ 8 ]);
+                    $originalGifSize = filesize($_FILES["imagefile"]["tmp_name"]);
+                    $useZga = $originalGifSize > $zgaSize;
+                }
+                if ( !$useZga  ) {
                     unset($sizes[ 8 ]);
                     unset($files[ 8 ]);
                     unset($fileNames[ 8 ]);
